Memo's inserting ~working

+++ app/src/main/java/com/salach/privatescheduler/db/daos/ChoreDao.kt
@@ -16,7 +16,7 @@ interface ChoreDao {
 //    fun getAllFromNote(noteId: Int): Flow<List<Chore>>
 
     @Query("SELECT * FROM Chore WHERE id = :id")
-    suspend fun getById(id: Int): Chore
+    suspend fun getById(id: Long): Chore
 
     @Insert
     suspend fun insertAll(vararg chores: Chore)
